backend/pyproject.toml
[project]
name = "boat-backend"
version = "3.0.0"
description = "BoAT — Boğaziçi University Annotation Tool (FastAPI backend)"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.30",
    "alembic>=1.14",
    "pydantic-settings>=2.7",
    "python-jose[cryptography]>=3.3",
    "bcrypt>=4.2",
    "python-multipart>=0.0.18",
    "email-validator>=2.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "httpx>=0.28",
    "aiosqlite>=0.21",
    "ruff>=0.9",
]

[tool.setuptools.packages.find]
include = ["app*"]

[tool.ruff]
target-version = "py312"
line-length = 99

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM", "RUF"]
ignore = ["B008"]  # Depends() in FastAPI route signatures is idiomatic

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
